added test data for Parker AIF
