+++ test/PopulationAIF_DCE/data/ParkerAIF.csv
@@ -1,0 +1,2 @@
+Parameter,A1,sigma1,T1,A2,sigma2,T2,alpha,beta,s,tau
+value,0.809,0.0563,0.17046,0.33,0.132,0.365,1.05,0.1685,38.078,0.483
