--- test/PopulationAIF_DCE/data/ParkerAIF.txt
@@ -1,10 +1,0 @@
-A1	48.54
-Sigma1	3.378
-T1	10.2276
-A2	19.8
-sigma2	7.92
-T2	21.9
-alpha	1.05
-beta	0.0028
-s	0.6346
-tau	28.98
